finish pb2mnn
