+++ deploy/utils.py
@@ -1,0 +1,164 @@
+# coding: utf-8
+
+import numpy as np
+import tensorflow as tf
+import random
+import colorsys
+import cv2
+
+def iou_calc1(boxes1, boxes2):
+    """
+    :param boxes1: boxes1和boxes2的shape可以不相同，但是需要满足广播机制
+    :param boxes2: 且需要保证最后一维为坐标维，以及坐标的存储结构为(xmin, ymin, xmax, ymax)
+    :return: 返回boxes1和boxes2的IOU，IOU的shape为boxes1和boxes2广播后的shape[:-1]
+    """
+    boxes1 = np.array(boxes1)
+    boxes2 = np.array(boxes2)
+
+    boxes1_area = (boxes1[..., 2] - boxes1[..., 0]) * (boxes1[..., 3] - boxes1[..., 1])
+    boxes2_area = (boxes2[..., 2] - boxes2[..., 0]) * (boxes2[..., 3] - boxes2[..., 1])
+
+    # 计算出boxes1和boxes2相交部分的左上角坐标、右下角坐标
+    left_up = np.maximum(boxes1[..., :2], boxes2[..., :2])
+    right_down = np.minimum(boxes1[..., 2:], boxes2[..., 2:])
+
+    # 计算出boxes1和boxes2相交部分的宽、高
+    # 因为两个boxes没有交集时，(right_down - left_up) < 0，所以maximum可以保证当两个boxes没有交集时，它们之间的iou为0
+    inter_section = np.maximum(right_down - left_up, 0.0)
+    inter_area = inter_section[..., 0] * inter_section[..., 1]
+    union_area = boxes1_area + boxes2_area - inter_area
+    IOU = 1.0 * inter_area / union_area
+    return IOU
+
+def nms(bboxes, score_threshold, iou_threshold, sigma=0.3, method='nms'):
+    """
+    :param bboxes:
+    假设有N个bbox的score大于score_threshold，那么bboxes的shape为(N, 6)，存储格式为(xmin, ymin, xmax, ymax, score, class)
+    其中(xmin, ymin, xmax, ymax)的大小都是相对于输入原图的，score = conf * prob，class是bbox所属类别的索引号
+    :return: best_bboxes
+    假设NMS后剩下N个bbox，那么best_bboxes的shape为(N, 6)，存储格式为(xmin, ymin, xmax, ymax, score, class)
+    其中(xmin, ymin, xmax, ymax)的大小都是相对于输入原图的，score = conf * prob，class是bbox所属类别的索引号
+    """
+    classes_in_img = list(set(bboxes[:, 5]))
+    best_bboxes = []
+
+    for cls in classes_in_img:
+        cls_mask = (bboxes[:, 5] == cls)
+        cls_bboxes = bboxes[cls_mask]
+        while len(cls_bboxes) > 0:
+            max_ind = np.argmax(cls_bboxes[:, 4])
+            best_bbox = cls_bboxes[max_ind]
+            best_bboxes.append(best_bbox)
+            cls_bboxes = np.concatenate([cls_bboxes[: max_ind], cls_bboxes[max_ind + 1:]])
+            iou = iou_calc1(best_bbox[np.newaxis, :4], cls_bboxes[:, :4])
+            assert method in ['nms', 'soft-nms']
+            weight = np.ones((len(iou),), dtype=np.float32)
+            if method == 'nms':
+                iou_mask = iou > iou_threshold
+                weight[iou_mask] = 0.0
+            if method == 'soft-nms':
+                weight = np.exp(-(1.0 * iou ** 2 / sigma))
+            cls_bboxes[:, 4] = cls_bboxes[:, 4] * weight
+            score_mask = cls_bboxes[:, 4] > score_threshold
+            cls_bboxes = cls_bboxes[score_mask]
+    return best_bboxes
+
+def preprocess(image, bboxes, target_shape, correct_box=True, keepratio=True):
+    """
+    RGB转换 -> resize(resize不改变原图的高宽比) -> normalize
+    并可以选择是否校正bbox
+    :param image_org: 要处理的图像
+    :param target_shape: 对图像处理后，期望得到的图像shape，存储格式为(h, w)
+    :return: 处理之后的图像，shape为target_shape
+    """
+    h_target, w_target = target_shape
+    h_org, w_org, _ = image.shape
+
+    image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB).astype(np.float32)
+    if not keepratio:
+        ratio_w = w_target / w_org
+        ratio_h = h_target / h_org
+        image = cv2.resize(image, target_shape, interpolation=cv2.INTER_LINEAR) / 255.0
+        if correct_box:
+            bboxes[:, [0, 2]] = bboxes[:, [0, 2]] * ratio_w
+            bboxes[:, [1, 3]] = bboxes[:, [1, 3]] * ratio_h
+            return image, bboxes
+        return image
+    resize_ratio = min(1.0 * w_target / w_org, 1.0 * h_target / h_org)
+    resize_w = int(resize_ratio * w_org)
+    resize_h = int(resize_ratio * h_org)
+    image_resized = cv2.resize(image, (resize_w, resize_h))
+    image_paded = np.full((h_target, w_target, 3), 128.0)
+    dw = int((w_target - resize_w) / 2)
+    dh = int((h_target - resize_h) / 2)
+    image_paded[dh:resize_h + dh, dw:resize_w + dw, :] = image_resized
+    image = image_paded / 255.0
+
+    if correct_box:
+        bboxes[:, [0, 2]] = bboxes[:, [0, 2]] * resize_ratio + dw
+        bboxes[:, [1, 3]] = bboxes[:, [1, 3]] * resize_ratio + dh
+        return image, bboxes
+    return image
+def postprocess(pred_bbox, test_input_size, org_img_shape):
+    conf_thres = 0.4
+    pred_bbox = np.array(pred_bbox)
+    pred_coor = pred_bbox[:, 0:4]
+    pred_conf = pred_bbox[:, 4]
+    pred_prob = pred_bbox[:, 5:]
+    org_h, org_w = org_img_shape
+    resize_ratio = min(1.0 * test_input_size / org_w, 1.0 * test_input_size / org_h)
+    dw = (test_input_size - resize_ratio * org_w) / 2
+    dh = (test_input_size - resize_ratio * org_h) / 2
+    pred_coor[:, 0::2] = 1.0 * (pred_coor[:, 0::2] - dw) / resize_ratio
+    pred_coor[:, 1::2] = 1.0 * (pred_coor[:, 1::2] - dh) / resize_ratio
+
+    pred_coor = np.concatenate([np.maximum(pred_coor[:, :2], [0, 0]),
+                                np.minimum(pred_coor[:, 2:], [org_w - 1, org_h - 1])], axis=-1)
+    invalid_mask = np.logical_or((pred_coor[:, 0] > pred_coor[:, 2]), (pred_coor[:, 1] > pred_coor[:, 3]))
+    pred_coor[invalid_mask] = 0
+    bboxes_scale = np.sqrt(np.multiply.reduce(pred_coor[:, 2:4] - pred_coor[:, 0:2], axis=-1))
+    valid_scale = (0, np.inf)
+    scale_mask = np.logical_and((valid_scale[0] < bboxes_scale), (bboxes_scale < valid_scale[1]))
+
+    classes = np.argmax(pred_prob, axis=-1)
+    scores = pred_conf * pred_prob[np.arange(len(pred_coor)), classes]
+    score_mask = scores > conf_thres
+    mask = np.logical_and(scale_mask, score_mask)
+
+    coors = pred_coor[mask]
+    scores = scores[mask]
+    classes = classes[mask]
+    bboxes = np.concatenate([coors, scores[:, np.newaxis], classes[:, np.newaxis]], axis=-1)
+    bboxes = nms(bboxes, conf_thres, 0.45, method='nms')
+    return bboxes
+
+
+def draw_bbox(original_image, bboxes, classes):
+    """
+    :param original_image: 检测的原始图片，shape为(org_h, org_w, 3)
+    :param bboxes: shape为(N, 6)，存储格式为(xmin, ymin, xmax, ymax, score, class)
+    其中(xmin, ymin, xmax, ymax)的大小都是相对于输入原图的，score = conf * prob，class是bbox所属类别的索引号
+    :return: None
+    """
+    num_classes = len(classes)
+    hsv_tuples = [(1.0 * x / num_classes, 1., 1.) for x in range(num_classes)]
+    colors = list(map(lambda x: colorsys.hsv_to_rgb(*x), hsv_tuples))
+    colors = list(map(lambda x: (int(x[0] * 255), int(x[1] * 255), int(x[2] * 255)), colors))
+    random.seed(0)
+    random.shuffle(colors)
+    random.seed(None)
+
+    image_h, image_w, _ = original_image.shape
+    for i, bbox in enumerate(bboxes):
+        coor = np.array(bbox[:4], dtype=np.int32)
+        score = bbox[4]
+        class_ind = int(bbox[5])
+        bbox_color = colors[class_ind]
+        bbox_thick = int(1.0 * (image_h + image_w) / 600)
+        cv2.rectangle(original_image, (coor[0], coor[1]), (coor[2], coor[3]), bbox_color, bbox_thick)
+
+        bbox_mess = '%s: %.3f' % (classes[class_ind], score)
+        text_loc = (int(coor[0]), int(coor[1] + 5) if coor[1] < 20 else int(coor[1] - 5))
+        cv2.putText(original_image, bbox_mess, text_loc, cv2.FONT_HERSHEY_SIMPLEX,
+                    1e-3 * image_h, (255, 255, 255), bbox_thick // 3)
+    return original_image
