+++ deploy/yolov3.py
@@ -2,13 +2,14 @@
 
 import numpy as np
 import tensorflow as tf
-from port2tf.layers import *
-from port2tf.MobilenetV2 import MobilenetV2,MobilenetV2_dynamic
+from layers import *
+from MobilenetV2 import MobilenetV2,MobilenetV2_dynamic
 
 class YOLOV3(object):
-    def __init__(self, training):
+    def __init__(self, training,numcls=20):
         self.__training = training
-        self.__num_classes = 20
+        self.__num_classes = numcls
+        self.__strides=[8,16,32]
     def build_nework(self, input_data, val_reuse=False,gt_per_grid=3):
         """
         :param input_data: shape为(batch_size, input_size, input_size, 3)
@@ -106,20 +107,20 @@ class YOLOV3(object):
             feature_map_s, feature_map_m, feature_map_l = MobilenetV2_dynamic(input_data, self.__training,statedict)
 
             conv = convolutional(name='conv0', input_data=feature_map_l, filters_shape=(1, 1, 1280, 512),
-                                 training=self.__training,statedict=statedict[19])
-            conv = separable_conv(name='conv1', input_data=conv, input_c=512, output_c=1024, training=self.__training,statedict=statedict[20])
+                                 training=self.__training,statedict=statedict['headslarge.conv0'])
+            conv = separable_conv(name='conv1', input_data=conv, input_c=512, output_c=1024, training=self.__training,statedict=statedict['headslarge.conv1'])
             conv = convolutional(name='conv2', input_data=conv, filters_shape=(1, 1, 1024, 512),
-                                 training=self.__training,statedict=statedict[21])
-            conv = separable_conv(name='conv3', input_data=conv, input_c=512, output_c=1024, training=self.__training,statedict=statedict[22])
+                                 training=self.__training,statedict=statedict['headslarge.conv2'])
+            conv = separable_conv(name='conv3', input_data=conv, input_c=512, output_c=1024, training=self.__training,statedict=statedict['headslarge.conv3'])
             conv = convolutional(name='conv4', input_data=conv, filters_shape=(1, 1, 1024, 512),
-                                 training=self.__training,statedict=statedict[23])
+                                 training=self.__training,statedict=statedict['headslarge.conv4'])
 
             # ----------**********---------- Detection branch of large object ----------**********----------
             conv_lbbox = separable_conv(name='conv5', input_data=conv, input_c=512, output_c=1024,
-                                        training=self.__training,statedict=statedict[24])
+                                        training=self.__training,statedict=statedict['detlarge.conv5'])
             conv_lbbox = convolutional(name='conv6', input_data=conv_lbbox,
                                        filters_shape=(1, 1, 1024, gt_per_grid * (self.__num_classes + 5)),
-                                       training=self.__training, downsample=False, activate=False, bn=False,statedict=statedict[25])
+                                       training=self.__training, downsample=False, activate=False, bn=False,statedict=statedict['detlarge.conv6'])
             pred_lbbox = decode_validate(name='pred_lbbox', conv_output=conv_lbbox,
                                     num_classes=self.__num_classes, stride=self.__strides[2], shape=inputsize // 32,
                                     gt_pergrid=gt_per_grid)
@@ -127,26 +128,26 @@ class YOLOV3(object):
 
             # ----------**********---------- up sample and merge features map ----------**********----------
             conv = convolutional(name='conv7', input_data=conv, filters_shape=(1, 1, 512, 256),
-                                 training=self.__training,statedict=statedict[26])
+                                 training=self.__training,statedict=statedict['mergelarge.conv7'])
             conv = upsample_decode(name='upsample0', input_data=conv,shape1=inputsize//32,shape2=inputsize//32)
             conv = route(name='route0', previous_output=feature_map_m, current_output=conv)
             # ----------**********---------- up sample and merge features map ----------**********----------
 
             conv = convolutional('conv8', input_data=conv, filters_shape=(1, 1, 96 + 256, 256),
-                                 training=self.__training,statedict=statedict[27])
-            conv = separable_conv('conv9', input_data=conv, input_c=256, output_c=512, training=self.__training,statedict=statedict[28])
+                                 training=self.__training,statedict=statedict['headsmid.conv8'])
+            conv = separable_conv('conv9', input_data=conv, input_c=256, output_c=512, training=self.__training,statedict=statedict['headsmid.conv9'])
             conv = convolutional('conv10', input_data=conv, filters_shape=(1, 1, 512, 256),
-                                 training=self.__training,statedict=statedict[29])
-            conv = separable_conv('conv11', input_data=conv, input_c=256, output_c=512, training=self.__training,statedict=statedict[30])
+                                 training=self.__training,statedict=statedict['headsmid.conv10'])
+            conv = separable_conv('conv11', input_data=conv, input_c=256, output_c=512, training=self.__training,statedict=statedict['headsmid.conv11'])
             conv = convolutional('conv12', input_data=conv, filters_shape=(1, 1, 512, 256),
-                                 training=self.__training,statedict=statedict[31])
+                                 training=self.__training,statedict=statedict['headsmid.conv12'])
 
             # ----------**********---------- Detection branch of middle object ----------**********----------
             conv_mbbox = separable_conv(name='conv13', input_data=conv, input_c=256, output_c=512,
-                                        training=self.__training,statedict=statedict[32])
+                                        training=self.__training,statedict=statedict['detmid.conv13'])
             conv_mbbox = convolutional(name='conv14', input_data=conv_mbbox,
                                        filters_shape=(1, 1, 512, gt_per_grid * (self.__num_classes + 5)),
-                                       training=self.__training, downsample=False, activate=False, bn=False,statedict=statedict[33])
+                                       training=self.__training, downsample=False, activate=False, bn=False,statedict=statedict['detmid.conv14'])
             pred_mbbox = decode_validate(name='pred_mbbox', conv_output=conv_mbbox,
                                     num_classes=self.__num_classes, stride=self.__strides[1], shape=inputsize // 16,
                                     gt_pergrid=gt_per_grid)
@@ -154,26 +155,26 @@ class YOLOV3(object):
 
             # ----------**********---------- up sample and merge features map ----------**********----------
             conv = convolutional(name='conv15', input_data=conv, filters_shape=(1, 1, 256, 128),
-                                 training=self.__training,statedict=statedict[34])
+                                 training=self.__training,statedict=statedict['mergemid.conv15'])
             conv = upsample_decode(name='upsample1', input_data=conv,shape1=inputsize//16,shape2=inputsize//16)
             conv = route(name='route1', previous_output=feature_map_s, current_output=conv)
             # ----------**********---------- up sample and merge features map ----------**********----------
 
             conv = convolutional(name='conv16', input_data=conv, filters_shape=(1, 1, 32 + 128, 128),
-                                 training=self.__training,statedict=statedict[35])
-            conv = separable_conv(name='conv17', input_data=conv, input_c=128, output_c=256, training=self.__training,statedict=statedict[36])
+                                 training=self.__training,statedict=statedict['headsmall.conv16'])
+            conv = separable_conv(name='conv17', input_data=conv, input_c=128, output_c=256, training=self.__training,statedict=statedict['headsmall.conv17'])
             conv = convolutional(name='conv18', input_data=conv, filters_shape=(1, 1, 256, 128),
-                                 training=self.__training,statedict=statedict[37])
-            conv = separable_conv(name='conv19', input_data=conv, input_c=128, output_c=256, training=self.__training,statedict=statedict[38])
+                                 training=self.__training,statedict=statedict['headsmall.conv18'])
+            conv = separable_conv(name='conv19', input_data=conv, input_c=128, output_c=256, training=self.__training,statedict=statedict['headsmall.conv19'])
             conv = convolutional(name='conv20', input_data=conv, filters_shape=(1, 1, 256, 128),
-                                 training=self.__training,statedict=statedict[39])
+                                 training=self.__training,statedict=statedict['headsmall.conv20'])
 
             # ----------**********---------- Detection branch of small object ----------**********----------
             conv_sbbox = separable_conv(name='conv21', input_data=conv, input_c=128, output_c=256,
-                                        training=self.__training,statedict=statedict[40])
+                                        training=self.__training,statedict=statedict['detsmall.conv21'])
             conv_sbbox = convolutional(name='conv22', input_data=conv_sbbox,
                                        filters_shape=(1, 1, 256, gt_per_grid * (self.__num_classes + 5)),
-                                       training=self.__training, downsample=False, activate=False, bn=False,statedict=statedict[41])
+                                       training=self.__training, downsample=False, activate=False, bn=False,statedict=statedict['detsmall.conv22'])
             pred_sbbox = decode_validate(name='pred_sbbox', conv_output=conv_sbbox,
                                     num_classes=self.__num_classes, stride=self.__strides[0], shape=inputsize // 8,
                                     gt_pergrid=gt_per_grid)
